requirements.txt
MouseInfo==0.1.3
mss==6.1.0
numpy==1.23.3
opencv-python==4.6.0.66
Pillow==9.2.0
PyAutoGUI==0.9.53
pygame==2.1.2
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
PySDL2==0.9.14
python3-xlib==0.15
pytweening==1.0.4
